Implemented fallback Git support with isomorphic-git integration (old approach
- Introduced a routing layer to switch between native dugite and isomorphic-git based on user preferences.
- Updated UI components to reflect Git tool status and provide user feedback on the current mode.
- Implemented necessary changes in various modules to support the new Git functionality, including command handling and status reporting.
