+++ package.json
@@ -1101,6 +1101,7 @@
         "hog-features": "^1.0.0",
         "i": "^0.3.7",
         "immutability-helper": "^3.1.1",
+        "isomorphic-git": "^1.27.1",
         "jimp": "^0.22.12",
         "md5": "^2.3.0",
         "meilisearch": "^0.38.0",
